+++ packages/ar_quick_look/lib/ar_quick_look.dart
@@ -3,6 +3,7 @@ library;
 
 export 'src/ar_quick_look.dart'
     show
+        AlreadyPresentingException,
         ArQuickLook,
         ArQuickLookException,
         FileNotFoundException,